requirements.txt
# Core FastAPI dependencies
fastapi>=0.68.0
uvicorn>=0.15.0
pydantic>=1.8.0

# Google Gemini AI
google-generativeai==0.3.2

# HTTP client (needed for testing and some integrations)
httpx>=0.24.1

# CORS support (if needed for frontend)
python-multipart>=0.0.5 